refactor(ts): split execute_node.ts into program/command_dispatch/classify_argv (mirror #146)
