+++ typescript/packages/core/src/workspace/node/command_dispatch.ts
@@ -1,0 +1,401 @@
+// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+
+import type { CommandHistory } from '../../commands/config.ts'
+import { type ByteSource, IOResult, materialize } from '../../io/types.ts'
+import { classifyArgvBySpec } from './classify_argv.ts'
+import type { Resource } from '../../resource/base.ts'
+import type { CallStack } from '../../shell/call_stack.ts'
+import {
+  ProcessSubDirection,
+  getCommandName,
+  getParts,
+  getProcessSubDirection,
+  getText,
+} from '../../shell/helpers.ts'
+import type { PyodideRuntime } from '../executor/python/runtime.ts'
+import type { JobTable } from '../../shell/job_table.ts'
+import { NodeType as NT, ShellBuiltin as SB } from '../../shell/types.ts'
+import { PathSpec } from '../../types.ts'
+import { classifyBarePath, classifyParts } from '../expand/classify.ts'
+import { type ExecuteFn, expandNode } from '../expand/node.ts'
+import { expandParts } from '../expand/parts.ts'
+import type { TSNodeLike } from '../expand/variable.ts'
+import { handleCommand } from '../executor/command.ts'
+import { runWithTimeout } from '../../commands/builtin/utils/safeguard.ts'
+import { resolveSafeguard } from '../../commands/safeguard.ts'
+import { BreakSignal, ContinueSignal } from '../executor/control.ts'
+import type { DispatchFn } from '../executor/cross_mount.ts'
+import {
+  handleBash,
+  handleCd,
+  handleEcho,
+  handleEval,
+  handleExport,
+  handleLocal,
+  handleMan,
+  handlePrintenv,
+  handlePrintf,
+  handleRead,
+  handleReturn,
+  handleSet,
+  handleShift,
+  handleSleep,
+  handleSource,
+  handleTest,
+  handleTrap,
+  handleUnset,
+  handleWhoami,
+} from '../executor/builtins/index.ts'
+import type { MountRegistry } from '../mount/registry.ts'
+import type { Session } from '../session/session.ts'
+import { ExecutionNode } from '../types.ts'
+import { resolveGlobs } from './resolve_globs.ts'
+
+type Result = [ByteSource | null, IOResult, ExecutionNode]
+
+const UNSUPPORTED_BUILTINS: ReadonlySet<string> = new Set([
+  'bg',
+  'disown',
+  'exec',
+  'complete',
+  'compgen',
+  'ulimit',
+])
+
+export async function executeCommand(
+  recurse: (
+    n: TSNodeLike,
+    s: Session,
+    i: ByteSource | null,
+    cs: CallStack | null,
+  ) => Promise<Result>,
+  dispatch: DispatchFn,
+  registry: MountRegistry,
+  executeFn: ExecuteFn,
+  node: TSNodeLike,
+  session: Session,
+  stdinIn: ByteSource | null,
+  callStack: CallStack | null,
+  jobTable: JobTable | null,
+  ensureOpen?: (resource: Resource) => Promise<void>,
+  unmount?: (prefix: string) => Promise<void>,
+  pythonRuntime?: PyodideRuntime,
+  history?: CommandHistory,
+  signal?: AbortSignal,
+): Promise<Result> {
+  const name = getCommandName(node)
+  const rawParts = getParts(node)
+
+  const prefixAssignments: [string, string][] = []
+  const nonPrefixParts: TSNodeLike[] = []
+  let sawCommandName = false
+  for (const p of rawParts) {
+    if (!sawCommandName && p.type === NT.VARIABLE_ASSIGNMENT) {
+      const atext = getText(p)
+      const eq = atext.indexOf('=')
+      if (eq >= 0) {
+        const key = atext.slice(0, eq)
+        const rawVal = atext.slice(eq + 1)
+        const valNodes = p.namedChildren.filter((c) => c.type !== NT.VARIABLE_NAME)
+        const firstVal = valNodes[0]
+        const v =
+          firstVal !== undefined
+            ? await expandNode(firstVal, session, executeFn, callStack)
+            : rawVal
+        prefixAssignments.push([key, v])
+      }
+      continue
+    }
+    if (p.type === NT.COMMAND_NAME) sawCommandName = true
+    nonPrefixParts.push(p)
+  }
+
+  for (const [k] of prefixAssignments) {
+    if (session.readonlyVars.has(k)) {
+      const err = new TextEncoder().encode(`bash: ${k}: readonly variable\n`)
+      return [
+        null,
+        new IOResult({ exitCode: 1, stderr: err }),
+        new ExecutionNode({ command: name !== '' ? name : k, exitCode: 1, stderr: err }),
+      ]
+    }
+  }
+
+  if (prefixAssignments.length > 0 && name === '') {
+    for (const [k, v] of prefixAssignments) session.env[k] = v
+    const cmdLabel = prefixAssignments.map(([k, v]) => `${k}=${v}`).join(' ')
+    return [null, new IOResult(), new ExecutionNode({ command: cmdLabel, exitCode: 0 })]
+  }
+
+  const isFunctionCall = name !== '' && session.functions[name] !== undefined
+  const savedEnvOverrides: Record<string, string | null> = {}
+  for (const [k, v] of prefixAssignments) {
+    if (!isFunctionCall) savedEnvOverrides[k] = k in session.env ? (session.env[k] ?? null) : null
+    session.env[k] = v
+  }
+
+  const resolved = name !== '' ? resolveSafeguard(name) : null
+  const timeout = resolved !== null ? resolved.timeoutSeconds : null
+  try {
+    return await runWithTimeout(
+      runCommandBody(
+        recurse,
+        dispatch,
+        registry,
+        executeFn,
+        node,
+        nonPrefixParts,
+        name,
+        session,
+        stdinIn,
+        callStack,
+        jobTable,
+        ensureOpen,
+        unmount,
+        pythonRuntime,
+        history,
+        signal,
+      ),
+      timeout,
+      name !== '' ? name : '?',
+    )
+  } finally {
+    for (const [k, prev] of Object.entries(savedEnvOverrides)) {
+      if (prev === null) {
+        // eslint-disable-next-line @typescript-eslint/no-dynamic-delete
+        delete session.env[k]
+      } else {
+        session.env[k] = prev
+      }
+    }
+  }
+}
+
+async function runCommandBody(
+  recurse: (
+    n: TSNodeLike,
+    s: Session,
+    i: ByteSource | null,
+    cs: CallStack | null,
+  ) => Promise<Result>,
+  dispatch: DispatchFn,
+  registry: MountRegistry,
+  executeFn: ExecuteFn,
+  node: TSNodeLike,
+  parts: TSNodeLike[],
+  name: string,
+  session: Session,
+  stdinIn: ByteSource | null,
+  callStack: CallStack | null,
+  jobTable: JobTable | null,
+  ensureOpen?: (resource: Resource) => Promise<void>,
+  unmount?: (prefix: string) => Promise<void>,
+  pythonRuntime?: PyodideRuntime,
+  history?: CommandHistory,
+  signal?: AbortSignal,
+): Promise<Result> {
+  let stdin = stdinIn
+
+  for (const child of node.namedChildren) {
+    if (child.type === NT.HERESTRING_REDIRECT) {
+      for (const sc of child.namedChildren) {
+        const content = await expandNode(sc, session, executeFn, callStack)
+        stdin = new TextEncoder().encode(`${content}\n`)
+        break
+      }
+    }
+  }
+
+  const procSubParts: Uint8Array[] = []
+  const cleanParts: TSNodeLike[] = []
+  for (const p of parts) {
+    if (p.type === NT.PROCESS_SUBSTITUTION) {
+      if (getProcessSubDirection(p) === ProcessSubDirection.OUTPUT) {
+        const err = new TextEncoder().encode('mirage: unsupported: process substitution >(...)\n')
+        return [
+          null,
+          new IOResult({ exitCode: 2, stderr: err }),
+          new ExecutionNode({
+            command: name === '' ? 'process_sub' : name,
+            exitCode: 2,
+            stderr: err,
+          }),
+        ]
+      }
+      const innerCmds = p.namedChildren.filter((c) => c.type === NT.COMMAND)
+      const innerFirst = innerCmds[0]
+      if (innerFirst !== undefined) {
+        const io = await executeFn(getText(innerFirst), { sessionId: session.sessionId })
+        procSubParts.push(await materialize(io.stdout))
+      }
+      continue
+    }
+    cleanParts.push(p)
+  }
+  if (procSubParts.length > 0 && stdin === null) {
+    let total = 0
+    for (const c of procSubParts) total += c.byteLength
+    const merged = new Uint8Array(total)
+    let off = 0
+    for (const c of procSubParts) {
+      merged.set(c, off)
+      off += c.byteLength
+    }
+    stdin = merged
+  }
+
+  const expanded = await expandParts(cleanParts, session, executeFn, callStack)
+
+  let textArgs: ReadonlySet<string> | null = null
+  let pathArgs: ReadonlySet<string> | null = null
+  const cwdMount = registry.mountFor(session.cwd)
+  const spec = cwdMount !== null ? cwdMount.specFor(name) : null
+  if (spec !== null) {
+    const [textSet, pathSet] = classifyArgvBySpec(spec, expanded.slice(1))
+    textArgs = textSet.size > 0 ? textSet : null
+    pathArgs = pathSet.size > 0 ? pathSet : null
+  }
+
+  const classified = classifyParts(expanded, registry, session.cwd, textArgs, pathArgs)
+  const resolved = await resolveGlobs(classified, registry, textArgs)
+  const finalExpanded = resolved.map((p) => (p instanceof PathSpec ? p.original : p))
+
+  // Unsupported bash builtins. Constructs the parser accepts but the
+  // executor cannot honor. Returning a clear error lets LLMs detect a
+  // capability gap instead of treating it as a missing binary.
+  if (UNSUPPORTED_BUILTINS.has(name)) {
+    const err = new TextEncoder().encode(`mirage: unsupported builtin: ${name}\n`)
+    return [
+      null,
+      new IOResult({ exitCode: 2, stderr: err }),
+      new ExecutionNode({ command: name, exitCode: 2, stderr: err }),
+    ]
+  }
+
+  // Shell builtins
+  if (name === SB.PWD) {
+    const out = new TextEncoder().encode(`${session.cwd}\n`)
+    return [out, new IOResult(), new ExecutionNode({ command: 'pwd', exitCode: 0 })]
+  }
+
+  if (name === SB.CD) {
+    let path: string | PathSpec = '/'
+    if (classified.length > 1) {
+      const raw = classified[1]
+      const rawStr = raw instanceof PathSpec ? raw.original : String(raw)
+      if (rawStr === '~') path = '/'
+      else if (raw instanceof PathSpec) path = raw
+      else if (rawStr.startsWith('/')) path = rawStr
+      else path = classifyBarePath(rawStr, registry, session.cwd)
+    }
+    return handleCd(dispatch, (p) => registry.isMountRoot(p), path, session)
+  }
+
+  if (name === SB.TRUE) {
+    return [null, new IOResult(), new ExecutionNode({ command: 'true', exitCode: 0 })]
+  }
+
+  if (name === SB.FALSE) {
+    return [
+      null,
+      new IOResult({ exitCode: 1 }),
+      new ExecutionNode({ command: 'false', exitCode: 1 }),
+    ]
+  }
+
+  if (name === SB.EVAL) return handleEval(executeFn, finalExpanded.slice(1), session)
+  if (name === SB.BASH || name === SB.SH) {
+    return handleBash(executeFn, finalExpanded.slice(1), session, stdin)
+  }
+  if (name === SB.EXPORT) return handleExport(finalExpanded.slice(1), session)
+  if (name === SB.UNSET) return handleUnset(finalExpanded.slice(1), session)
+  if (name === SB.LOCAL) return handleLocal(finalExpanded.slice(1), session)
+  if (name === SB.PRINTENV) {
+    return handlePrintenv(finalExpanded.length > 1 ? (finalExpanded[1] ?? null) : null, session)
+  }
+  if (name === SB.WHOAMI) return handleWhoami(session)
+  if (name === SB.MAN) return handleMan(finalExpanded.slice(1), session, registry)
+  if (name === SB.SET) return handleSet(finalExpanded.slice(1), session, callStack)
+  if (name === SB.SHIFT) {
+    const n = finalExpanded.length > 1 ? Number(finalExpanded[1]) : 1
+    return handleShift(Number.isFinite(n) ? n : 1, callStack, session)
+  }
+  if (name === SB.TRAP) return handleTrap(session)
+  if (name === SB.TEST || name === SB.BRACKET || name === SB.DOUBLE_BRACKET) {
+    return handleTest(dispatch, classified.slice(1), session)
+  }
+  if (name === SB.ECHO) {
+    const args = finalExpanded.slice(1)
+    const nFlag = args.includes('-n')
+    const eFlag = args.includes('-e')
+    return handleEcho(
+      args.filter((a) => a !== '-n' && a !== '-e'),
+      nFlag,
+      eFlag,
+    )
+  }
+  if (name === SB.PRINTF) return handlePrintf(finalExpanded.slice(1))
+  if (name === SB.SLEEP) return handleSleep(finalExpanded.slice(1), signal)
+  if (name === SB.READ) {
+    return handleRead(finalExpanded.slice(1), session, stdin)
+  }
+  if (name === SB.SOURCE || name === SB.DOT) {
+    const target = classified.length > 1 ? (classified[1] ?? '') : ''
+    return handleSource(dispatch, executeFn, target, session)
+  }
+  if (name === SB.RETURN) {
+    const n = finalExpanded.length > 1 ? Number(finalExpanded[1]) : 0
+    return handleReturn(Number.isFinite(n) ? n : 0)
+  }
+  if (name === SB.BREAK) throw new BreakSignal()
+  if (name === SB.CONTINUE) throw new ContinueSignal()
+
+  if (name === SB.XARGS) {
+    const stdinBytes = await materialize(stdin)
+    const inputArgs = new TextDecoder()
+      .decode(stdinBytes)
+      .split(/\s+/)
+      .filter((s) => s !== '')
+    const xargsCmd = finalExpanded[1] ?? 'echo'
+    const inner = `${xargsCmd} ${inputArgs.join(' ')}`
+    const io = await executeFn(inner, { sessionId: session.sessionId })
+    return [io.stdout, io, new ExecutionNode({ command: 'xargs', exitCode: io.exitCode })]
+  }
+
+  if (name === SB.TIMEOUT) {
+    if (finalExpanded.length >= 3) {
+      const innerCmd = finalExpanded.slice(2).join(' ')
+      const io = await executeFn(innerCmd, { sessionId: session.sessionId })
+      return [io.stdout, io, new ExecutionNode({ command: 'timeout', exitCode: io.exitCode })]
+    }
+    return [null, new IOResult(), new ExecutionNode({ command: 'timeout', exitCode: 0 })]
+  }
+
+  // Default: mount-dispatched command
+  return handleCommand(
+    recurse,
+    dispatch,
+    registry,
+    classified,
+    session,
+    stdin,
+    callStack,
+    jobTable,
+    ensureOpen,
+    unmount,
+    history,
+    pythonRuntime,
+  )
+}
