+++ typescript/packages/core/src/workspace/node/program.ts
@@ -1,0 +1,115 @@
+// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ========= Copyright 2026 @ Strukto.AI All Rights Reserved. =========
+
+import { asyncChain } from '../../io/stream.ts'
+import { type ByteSource, IOResult, materialize } from '../../io/types.ts'
+import type { CallStack } from '../../shell/call_stack.ts'
+import type { JobTable } from '../../shell/job_table.ts'
+import { ERREXIT_EXEMPT_TYPES, NodeType as NT } from '../../shell/types.ts'
+import type { TSNodeLike } from '../expand/variable.ts'
+import { handleBackground } from '../executor/jobs.ts'
+import type { Session } from '../session/session.ts'
+import { ExecutionNode } from '../types.ts'
+
+type Result = [ByteSource | null, IOResult, ExecutionNode]
+
+export async function executeProgram(
+  recurse: (
+    n: TSNodeLike,
+    s: Session,
+    i: ByteSource | null,
+    cs: CallStack | null,
+  ) => Promise<Result>,
+  node: TSNodeLike,
+  session: Session,
+  stdin: ByteSource | null,
+  callStack: CallStack | null,
+  jobTable: JobTable | null,
+  agentId: string,
+): Promise<Result> {
+  const children = node.children
+  const allStdout: ByteSource[] = []
+  let mergedIo = new IOResult()
+  let lastExec = new ExecutionNode({ command: '', exitCode: 0 })
+
+  let i = 0
+  while (i < children.length) {
+    const child = children[i]
+    if (child === undefined) {
+      i += 1
+      continue
+    }
+    if (child.isNamed !== true || child.type === NT.COMMENT) {
+      i += 1
+      continue
+    }
+    if (child.type === NT.ERROR) {
+      // ERROR nodes that contain only stray statement separators (`& ;`)
+      // are filtered out at parse-time by findSyntaxError, so anything
+      // reaching here is a recovered fragment we deliberately skip;
+      // structural errors would have raised before executeNode ran.
+      i += 1
+      continue
+    }
+
+    const next = children[i + 1]
+    const isBg = next?.type === NT.BACKGROUND
+
+    let stdout: ByteSource | null
+    let io: IOResult
+    if (isBg) {
+      const [bgStdout, bgIo, bgExec] = await handleBackground(
+        recurse,
+        child,
+        null,
+        session,
+        jobTable,
+        agentId,
+        stdin,
+        callStack,
+      )
+      stdout = bgStdout
+      io = bgIo
+      lastExec = bgExec
+      i += 2
+    } else {
+      const [s, ioResult, execNode] = await recurse(child, session, stdin, callStack)
+      stdout = await materialize(s)
+      ioResult.syncExitCode()
+      session.lastExitCode = ioResult.exitCode
+      io = ioResult
+      lastExec = execNode
+      i += 1
+    }
+
+    if (stdout !== null) allStdout.push(stdout)
+    mergedIo = await mergedIo.merge(io)
+
+    if (
+      io.exitCode !== 0 &&
+      session.shellOptions.errexit === true &&
+      !isBg &&
+      !ERREXIT_EXEMPT_TYPES.has(child.type)
+    ) {
+      mergedIo.exitCode = io.exitCode
+      break
+    }
+  }
+
+  if (allStdout.length === 1 && allStdout[0] !== undefined) {
+    return [allStdout[0], mergedIo, lastExec]
+  }
+  const combined = allStdout.length > 0 ? asyncChain(...allStdout) : null
+  return [combined, mergedIo, lastExec]
+}
